requirements.txt
absl-py==0.6.1
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.2
anaconda-project==0.8.2
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
arrow==0.4.2
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
astropy==3.0.5
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.0.2
bokeh==1.0.1
boto==2.49.0
Bottleneck==1.2.1
cachetools==3.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.4.0
conda==4.6.2
conda-build==3.16.2
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.4.2
cycler==0.10.0
Cython==0.29
cytoolz==0.9.0.1
dask==0.20.0
datashape==0.5.4
DateTime==4.3
decorator==4.3.0
defusedxml==0.5.0
distributed==1.24.0
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.10
Flask==1.0.2
Flask-Cors==3.0.6
gast==0.2.0
gevent==1.3.7
glob2==0.6
gmpy2==2.0.8
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
greenlet==0.4.15
grpcio==1.12.1
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.12.0
hyperlink==18.0.0
icalendar==4.0.3
idna==2.7
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==0.6
incremental==17.5.0
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.0
jupyterlab-launcher==0.13.1
jupyterlab-server==0.2.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==16.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.25.0
locket==0.2.0
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
navigator-updater==0.2.1
nb-conda==2.2.1
nb-conda-kernels==2.2.0
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nltk==3.3
nose==1.3.7
notebook==5.7.0
numba==0.40.0
numexpr==2.6.8
numpy==1.15.4
numpydoc==0.8.0
oauthlib==3.0.1
odo==0.5.1
olefile==0.46
openpyxl==2.5.9
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.4.2
pluggy==0.8.0
ply==3.11
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.8
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.2.0
pylint==2.1.1
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
pytest==3.10.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.1
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.2
qtconsole==4.4.2
QtPy==1.5.2
requests==2.20.0
requests-oauthlib==1.2.0
rope==0.11.0
rsa==4.0
ruamel-yaml==0.15.46
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.13
statsmodels==0.9.0
sympy==1.3
tables==3.4.4
tblib==1.3.2
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
Twisted==17.5.0
typed-ast==1.1.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.1.2
xlwings==0.14.0
xlwt==1.2.0
zict==0.1.3
zope.interface==4.6.0
